requirements.txt
Jinja2==3.0.2
html2image==2.0.1
newsapi-python==0.2.6
Pillow==8.4.0
PyQt5==5.15.4
requests
tqdm